Automate Sengernest using Gradle
Configure and setup Gradle to automate tasks for Sengernest.
