+++ build.gradle
@@ -28,7 +28,7 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+    mainClass.set("sengernest.Sengernest")
 }
 
 shadowJar {
