--- src/main/java/Sengernest/exceptions/EmptyTaskDescriptionException.java
@@ -1,6 +1,0 @@
-package Sengernest.exceptions;
-public class EmptyTaskDescriptionException extends Exception {
-    public EmptyTaskDescriptionException(String message) {
-        super(message);
-    }
-}
