--- src/main/java/Sengernest/exceptions/EmptyTaskListException.java
@@ -1,6 +1,0 @@
-package Sengernest.exceptions;
-public class EmptyTaskListException extends Exception {
-    public EmptyTaskListException(String message) {
-        super(message);
-    }
-}
