--- src/main/java/Sengernest/exceptions/InvalidDateFormatException.java
@@ -1,6 +1,0 @@
-package Sengernest.exceptions;
-public class InvalidDateFormatException extends Exception {
-    public InvalidDateFormatException(String message) {
-        super(message);
-    }
-}
