--- src/main/java/Sengernest/exceptions/InvalidTaskNumberException.java
@@ -1,6 +1,0 @@
-package Sengernest.exceptions;
-public class InvalidTaskNumberException extends Exception {
-    public InvalidTaskNumberException(String message) {
-        super(message);
-    }
-}
