--- src/main/java/Sengernest/exceptions/MarkFinishedTaskException.java
@@ -1,6 +1,0 @@
-package Sengernest.exceptions;
-public class MarkFinishedTaskException extends Exception {
-    public MarkFinishedTaskException(String message) {
-        super(message);
-    }
-}
