--- src/main/java/Sengernest/exceptions/MissingDateException.java
@@ -1,6 +1,0 @@
-package Sengernest.exceptions;
-public class MissingDateException extends Exception {
-    public MissingDateException(String message) {
-        super(message);
-    }
-}
