--- src/main/java/Sengernest/exceptions/MissingTaskNumberException.java
@@ -1,6 +1,0 @@
-package Sengernest.exceptions;
-public class MissingTaskNumberException extends Exception {
-    public MissingTaskNumberException(String message) {
-        super(message);
-    }
-}
