--- src/main/java/Sengernest/exceptions/UnknownCommandException.java
@@ -1,6 +1,0 @@
-package Sengernest.exceptions;
-public class UnknownCommandException extends Exception {
-    public UnknownCommandException(String message) {
-        super(message);
-    }
-}
