--- src/main/java/Sengernest/exceptions/UnmarkUnfinishedTaskException.java
@@ -1,6 +1,0 @@
-package Sengernest.exceptions;
-public class UnmarkUnfinishedTaskException extends Exception {
-    public UnmarkUnfinishedTaskException(String message) {
-        super(message);
-    }
-}
