--- src/main/java/Sengernest/tasks/ToDo.java
@@ -1,17 +1,0 @@
-package Sengernest.tasks;
-public class ToDo extends Task {
-
-    public ToDo(String tasking) {
-        super(tasking);
-    }
-
-    @Override
-    public String getTasking() {
-        return "[T]" + super.getTasking();
-    }
-
-    @Override
-    public String toFileFormat() {
-        return "T " + super.toFileFormat();
-    }
-}
