+++ src/main/java/sengernest/commands/PrintListCommand.java
@@ -1,0 +1,11 @@
+package sengernest.commands;
+import sengernest.storage.Storage;
+import sengernest.tasks.TaskList;
+import sengernest.ui.Ui;
+
+public class PrintListCommand extends Command {
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.printList(tasks);
+    }
+}
