+++ src/main/java/sengernest/exceptions/EmptyTaskDescriptionException.java
@@ -1,0 +1,6 @@
+package sengernest.exceptions;
+public class EmptyTaskDescriptionException extends Exception {
+    public EmptyTaskDescriptionException(String message) {
+        super(message);
+    }
+}
