+++ src/main/java/sengernest/exceptions/EmptyTaskListException.java
@@ -1,0 +1,6 @@
+package sengernest.exceptions;
+public class EmptyTaskListException extends Exception {
+    public EmptyTaskListException(String message) {
+        super(message);
+    }
+}
