+++ src/main/java/sengernest/exceptions/InvalidDateFormatException.java
@@ -1,0 +1,6 @@
+package sengernest.exceptions;
+public class InvalidDateFormatException extends Exception {
+    public InvalidDateFormatException(String message) {
+        super(message);
+    }
+}
