+++ src/main/java/sengernest/exceptions/InvalidTaskNumberException.java
@@ -1,0 +1,6 @@
+package sengernest.exceptions;
+public class InvalidTaskNumberException extends Exception {
+    public InvalidTaskNumberException(String message) {
+        super(message);
+    }
+}
