+++ src/main/java/sengernest/exceptions/MarkFinishedTaskException.java
@@ -1,0 +1,6 @@
+package sengernest.exceptions;
+public class MarkFinishedTaskException extends Exception {
+    public MarkFinishedTaskException(String message) {
+        super(message);
+    }
+}
