+++ src/main/java/sengernest/exceptions/MissingDateException.java
@@ -1,0 +1,6 @@
+package sengernest.exceptions;
+public class MissingDateException extends Exception {
+    public MissingDateException(String message) {
+        super(message);
+    }
+}
