+++ src/main/java/sengernest/exceptions/MissingTaskNumberException.java
@@ -1,0 +1,6 @@
+package sengernest.exceptions;
+public class MissingTaskNumberException extends Exception {
+    public MissingTaskNumberException(String message) {
+        super(message);
+    }
+}
