+++ src/main/java/sengernest/exceptions/UnknownCommandException.java
@@ -1,0 +1,6 @@
+package sengernest.exceptions;
+public class UnknownCommandException extends Exception {
+    public UnknownCommandException(String message) {
+        super(message);
+    }
+}
