+++ src/main/java/sengernest/exceptions/UnmarkUnfinishedTaskException.java
@@ -1,0 +1,6 @@
+package sengernest.exceptions;
+public class UnmarkUnfinishedTaskException extends Exception {
+    public UnmarkUnfinishedTaskException(String message) {
+        super(message);
+    }
+}
