+++ src/main/java/sengernest/tasks/Task.java
@@ -1,0 +1,47 @@
+package sengernest.tasks;
+public class Task {
+    private final String tasking;
+    private boolean finished;
+
+    public Task(String tasking) {
+        this.tasking = tasking;
+        this.finished = false;
+    }
+
+    public boolean isEmpty() {
+        return this.tasking.isEmpty();
+    }
+
+    public boolean isFinished() {
+        return this.finished;
+    }
+
+    public void finish() {
+        this.finished = true;
+    }
+
+    public void unfinish() {
+        this.finished = false;
+    }
+
+    public String getTasking() {
+        if (this.finished) {
+            return "[X] " + this.tasking;
+        } else {
+            return "[ ] " + this.tasking;
+        }
+    }
+
+    public String getTaskDescription() {
+        return this.tasking;
+    }
+
+    public String toFileFormat() {
+        if (this.isFinished()) {
+            return "| 1 | " + this.getTaskDescription(); 
+        } else {
+            return "| 0 | " + this.getTaskDescription() ; 
+        }
+    }
+}
+
