+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-Hello! I'm Sengernest
+Hello! I'm sengernest
 What can I do for you?: 
 UnknownCommandException: Invalid command! Valid commands: list, todo, Deadline, Event, mark, unmark, delete, bye.
